// lib/api/generated/models/athleteGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export interface AthleteGroup {
  created_at: string;
  id: string;
  member_user_ids?: string[];
  name: string;
  updated_at: string;
}