// lib/api/generated/models/assignmentKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type AssignmentKind = typeof AssignmentKind[keyof typeof AssignmentKind];


export const AssignmentKind = {
  none: 'none',
  all: 'all',
  users: 'users',
  groups: 'groups',
} as const;